+++ src/Public/Set-ShlinkUrl.ps1
@@ -133,17 +133,14 @@ function Set-ShlinkUrl {
 
             switch($PSBoundParameters.Keys) {
                 "AndroidLongUrl" {
-                    # TODO needs tests
                     $deviceLongUrls["android"] = $AndroidLongUrl
                     $Params["Body"]["deviceLongUrls"] = $deviceLongUrls
                 }
                 "IOSLongUrl" {
-                    # TODO needs tests
                     $deviceLongUrls["ios"] = $IOSLongUrl
                     $Params["Body"]["deviceLongUrls"] = $deviceLongUrls
                 }
                 "DesktopLongUrl" {
-                    # TODO needs tests
                     $deviceLongUrls["desktop"] = $DesktopLongUrl
                     $Params["Body"]["deviceLongUrls"] = $deviceLongUrls
                 }
@@ -175,6 +172,7 @@ function Set-ShlinkUrl {
                     }
                 }
                 "ValidateUrl" {
+                    Write-Warning 'validateUrl is deprecated since Shlink 3.5.0'
                     $Params["Body"]["validateUrl"] = $ValidateUrl
                 }
                 "ForwardQuery" {