+++ tests/invoke.tests.ps1
@@ -1,12 +1,12 @@
 $PesterConfig = New-PesterConfiguration -Hashtable @{
     Run = @{
         Path = @(
-            '{0}/Public/New-*.Acceptance.Tests.ps1' -f $PSScriptRoot
-            '{0}/Public/Get-*.Acceptance.Tests.ps1' -f $PSScriptRoot
-            '{0}/Public/Save-*.Acceptance.Tests.ps1' -f $PSScriptRoot
-            '{0}/Public/Set-ShlinkUrl.Acceptance.Tests.ps1' -f $PSScriptRoot
-            '{0}/Public/Set-ShlinkTag.Acceptance.Tests.ps1' -f $PSScriptRoot
-            '{0}/Public/Remove-*.Acceptance.Tests.ps1' -f $PSScriptRoot
+            '{0}/Public/New-*.Tests.ps1' -f $PSScriptRoot
+            '{0}/Public/Get-*.Tests.ps1' -f $PSScriptRoot
+            '{0}/Public/Save-*.Tests.ps1' -f $PSScriptRoot
+            '{0}/Public/Set-ShlinkUrl.Tests.ps1' -f $PSScriptRoot
+            '{0}/Public/Set-ShlinkTag.Tests.ps1' -f $PSScriptRoot
+            '{0}/Public/Remove-*.Tests.ps1' -f $PSScriptRoot
         )
         Throw = $true
         SkipRemainingOnFailure = 'Block'